update: File Record Service

+++ EmailApplication.Server/Data/FileAttachmentData.cs
@@ -2,7 +2,7 @@
 // Filename: FileAttachmentData.cs
 // Author: Aaron Thompson
 // Date Created: 3/10/2026
-// Last Updated: 4/29/2026
+// Last Updated: 4/30/2026
 //
 // Description: Data object for database access. UploaderID here exists to
 // to prevent users from swapping the FileID(s) on attachments on client side to
@@ -15,11 +15,11 @@ namespace EmailApplication.Server.Data {
 
 // VARIABLE(s)
 //------------------------------------------------------------------------------
-        public long FileID { get; set; } //PRIMARY KEY
+        public int FileID { get; set; } //PRIMARY KEY
         public string BucketKey { get; set; } //REFERENCES AWS S3 object key, format: attachments/{FileID}/{FileName}
         public int UploaderID { get; set; } //FOREIGN KEY, here for security reasons, see above for explanation.
         public string FileName { get; set; }
-        public int FileSize { get; set; }
+        public long FileSize { get; set; }
         public DateTime DateUploaded { get; set; }
         public DateTime DateLastReferenced { get; set;  }
         public int ReferenceCount { get; set; }
